+++ StripePaymentSheet/StripePaymentSheet/Source/Categories/String+Localized.swift
@@ -363,7 +363,6 @@ extension String.Localized {
     static var naver_pay_money_point: String {
         STPLocalizedString("Naver Pay Money/Point", "Naver Pay money or points funding source")
     }
-
     static var blik_confirm_payment: String {
         STPLocalizedString("Confirm the payment in your bank's app within %@ to complete the purchase.",
                            "Text for alert message when user needs to confirm payment in their banking app")