+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetFormFactory/PaymentSheetFormFactory.swift
@@ -295,7 +295,7 @@ class PaymentSheetFormFactory {
                 )
             case .mbWay, .bizum:
                 return makeContactInformationAndBillingAddressForm(phoneRequired: true)
-            case .cashApp, .payPal, .revolutPay, .amazonPay, .satispay, .twint:
+            case .cashApp, .payPal, .revolutPay, .amazonPay, .satispay, .twint, .krCard:
                 return makeContactInformationAndBillingAddressForm(
                     additionalElements: makeSetupMandateElements(for: paymentMethod)
                 )
@@ -307,7 +307,7 @@ class PaymentSheetFormFactory {
                 return makeAUBECSDebit()
             case .FPX:
                 return makeFPX()
-            case .kakaoPay, .netBanking, .weChatPay, .link, .cardPresent, .krCard, .unknown:
+            case .kakaoPay, .netBanking, .weChatPay, .link, .cardPresent, .krCard, .naverPay, .unknown:
                 return makeUnexpectedEmptyForm(for: paymentMethod)
             @unknown default:
                 return makeUnexpectedEmptyForm(for: paymentMethod)
@@ -343,6 +343,8 @@ class PaymentSheetFormFactory {
         case .twint:
             return [makeTwintMandate()]
         case .naverPay:
+            return [makeKoreanPaymentMethodMandate()]
+        case .krCard:
             return [makeKoreanPaymentMethodMandate()]
         default:
             return []