+++ CHANGELOG.md
@@ -7,7 +7,8 @@ MINOR
 
 ### PaymentSheet
 * [Added] Added support for Naver Pay.
-
+### PaymentSheet
+* [Added] Added support for Korean cards.
 ## 26.9.0 2026-08-31
 ### Payments
 * [Added] Added API bindings for Kakao Pay, SeQura, Korean cards, Naver Pay, and PAYCO.